# docs/openapi/OpenAPI Reference Implementation.json
{
  "swagger": "2.0",
  "info": {
    "title": "Identfy API",
    "description": "This api is Identfy",
    "license": {
      "name": "MIT License"
    },
    "version": "v1"
  },
  "schemes": [
    "https"
  ],
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "basic": {
      "type": "basic"
    }
  },
  "security": [
    {
      "Bearer": []
    },
    {
      "basic": []
    }
  ],
  "paths": {
...
    "/credentials/": {
      "post": {
        "operationId": "credentials",
        "description": "POST Credentials",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EbsiCredentialRequest"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Auth Token",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/CredentialResponse"
            }
          }
        },
        "tags": [
          "credentials"
        ]
      },
      "parameters": []
    },
...
    "/offers/{id}": {
      "get": {
        "operationId": "get_credential_offer_by_pk",
        "description": "GET Credential Offer Json",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Offer Identifier",
            "type": "string",
            "required": true
          },
          {
            "name": "pre-authorized_code",
            "in": "query",
            "description": "Preauthorize code",
            "type": "string"
          },
          {
            "name": "user_pin_required",
            "in": "query",
            "description": "User pin required",
            "type": "boolean"
          },
          {
            "name": "requested_vc_types",
            "in": "query",
            "description": "Credential types to include in the offer",
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/IssuanceOfferResponse"
            }
          }
        },
        "tags": [
          "get_credential_offer_by_pk"
        ]
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/presentations/external-data": {
      "post": {
        "operationId": "presentations_get_claims_validation",
        "description": "Post Claims Verification",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ExternalDataRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ClaimsVerification"
            }
          }
        },
        "tags": [
          "presentations"
        ]
      },
      "parameters": []
    },
    "/verify-flow": {
      "get": {
        "operationId": "verify_flow",
        "description": "Verify Flow",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "description": "Scope",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/VerifyFlow"
            }
          }
        },
        "tags": [
          "verify_flow"
        ]
      },
      "parameters": []
    },
...
  },
  "definitions": {
...
    "IssuanceOfferResponse": {
      "required": [
        "credential_offer"
      ],
      "type": "object",
      "properties": {
        "credential_offer": {
          "title": "Credential offer",
          "type": "object"
        },
        "pin": {
          "title": "Pin",
          "type": "integer"
        }
      }
    },
    "CredentialResponse": {
      "type": "object",
      "properties": {
        "format": {
          "title": "Format",
          "type": "string",
          "default": "jwt_vc",
          "maxLength": 255,
          "minLength": 1
        },
        "credential": {
          "title": "Credential",
          "type": "string",
          "maxLength": 1255,
          "minLength": 1
        },
        "acceptance_token": {
          "title": "Acceptance token",
          "type": "string",
          "maxLength": 1255,
          "minLength": 1
        },
        "c_nonce": {
          "title": "C nonce",
          "type": "string",
          "maxLength": 255,
          "minLength": 1
        },
        "c_nonce_expires_in": {
          "title": "C nonce expires in",
          "type": "integer"
        }
      }
    },
    "EbsiCredentialRequest": {
      "required": [
        "types",
        "proof"
      ],
      "type": "object",
      "properties": {
        "types": {
          "title": "Types",
          "description": "Credential types requested",
          "type": "object"
        },
        "format": {
          "title": "Format",
          "description": "Format of the returned credentials",
          "type": "string",
          "default": "jwt_vc",
          "maxLength": 255,
          "minLength": 1
        },
        "proof": {
          "title": "Proof",
          "description": "Proof object containing proof_type and jwt",
          "type": "object"
        }
      }
    },
...
    "PresentationDefinition": {
      "required": [
        "id",
        "input_descriptors"
      ],
      "type": "object",
      "properties": {
        "id": {
          "title": "Id",
          "type": "string",
          "minLength": 1
        },
        "format": {
          "title": "Format",
          "type": "object"
        },
        "input_descriptors": {
          "title": "Input descriptors",
          "type": "object"
        }
      }
    },
...
    "ExternalDataRequest": {
      "required": [
        "data"
      ],
      "type": "object",
      "properties": {
        "data": {
          "title": "Data",
          "type": "object"
        }
      }
    },
    "ClaimsVerification": {
      "required": [
        "verified"
      ],
      "type": "object",
      "properties": {
        "verified": {
          "title": "Verified",
          "type": "boolean"
        }
      }
    },
    "VerifyFlow": {
      "required": [
        "scope"
      ],
      "type": "object",
      "properties": {
        "scope": {
          "title": "Scope",
          "type": "string",
          "maxLength": 2000,
          "minLength": 1
        },
        "response_type": {
          "title": "Response Type",
          "type": "string",
          "enum": [
            "vp_token",
            "id_token"
          ]
        },
        "presentation_definition": {
          "$ref": "#/definitions/PresentationDefinition"
        }
      }
    },
...
  }
}